Delete: unused part delete
